feat: finalize core UI architecture and AWS AI state integration
- Overhauled DashboardScreen to match modern Figma UI with dynamic AI insight cards
- Implemented SettingsScreen for user CRUD operations and DB wiping (Developer Tools)
- Migrated analytics to dedicated HistoryScreen with full ledger
- Updated ExpenseDao with comprehensive Room queries (Today's Spend, Wipe Data)
- Integrated GamificationUiState into AddExpenseScreen to catch AWS Contextual Bandit actions
- Wired NavGraph to support full unidirectional data flow and seamless screen transitions

+++ app/src/main/java/com/finadapt/adaptivefinance/data/remote/FastApiInterface.kt
@@ -44,7 +44,7 @@ data class FeedbackRequest(
     @SerializedName("prediction_id")
     val predictionId: String,
     @SerializedName("reward")
-    val reward: Float
+    val reward: Int
 )
 
 //4. The exact FastAPI Endpoint
